services/agents/pyproject.toml
[project]
name = "infatium-agents"
version = "0.1.0"
description = "AI agents service for Infatium"
authors = [
    { name = "margosq", email = "slava1kvartovkin@gmail.com" }
]
dependencies = [
    "infatium-shared",
    "faststream[nats]>=0.6.4",
    "langchain>=0.3.0,<1.0.0",
    "langchain-core>=0.3.0,<1.0.0",
    "langchain-openai>=0.3.0,<1.0.0",
    "langchain-community>=0.3.31",
    "httpx>=0.28.1",
    "opentelemetry-instrumentation-anthropic>=0.30.0",
    "loguru>=0.7.3",
    "json-repair>=0.52.0",
    "aiocache>=0.12.2",
    "sentry-sdk[starlette,sqlalchemy,httpx,asyncio]>=2.0.0",
    "pydantic-settings>=2.11.0",
    "prometheus-client>=0.23.1",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp>=1.20.0",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["agents"]

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.15.1",
    "respx>=0.22.0",
    "ruff>=0.1.6",
]

[tool.rye.sources]
shared = { path = "../shared-python", editable = true }

[tool.rye.scripts]
lint = { chain = ["lint:ruff-check", "lint:ruff-format"] }
"lint:ruff-check" = "ruff check --fix agents/"
"lint:ruff-format" = "ruff format agents/"
test = "pytest tests/ -v"

[project.scripts]
infatium-agents = "agents.main:main"

[tool.ruff]
target-version = "py310"
line-length = 88
src = ["."]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "ARG001"]
ignore = ["E501", "B008", "C901", "W191"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["ARG001"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
